test(msa): keep the paged-varlen fp4 reference in float32
The kernel's fp4 x scale products never round through bf16, so the
bf16-rounded reference misses the 5e-2 tolerance on causal-boundary
blocks whose block-max reduces over near-cancelling dot products (16 of
40 seeds on an RTX 5080, worst 0.17; the checked-in seed fails at
0.083). Kept in float32 the kernel matches the reference exactly, the
same treatment _dequant_qk applies to the other fp4 tests.

AI-assisted (Claude Code).

+++ tests/msa_ops/test_proxy_fp4.py
@@ -285,13 +285,9 @@ def test_proxy_fp4_paged_varlen(Hq, Hkv, seqs_q):
     mkt = max(-(-s // BLK_KV) for s in seqs_k)
     assert out.shape == (Hq, mkt, total_q)
 
-    q_deq = (
-        _dequant_128x4(
-            q_fp4.reshape(-1, 64).cpu(), q_scale.cpu(), inv_q * inv_k, total_q * Hq
-        )
-        .reshape(total_q, Hq, 128)
-        .to(torch.bfloat16)
-    )
+    q_deq = _dequant_128x4(
+        q_fp4.reshape(-1, 64).cpu(), q_scale.cpu(), inv_q * inv_k, total_q * Hq
+    ).reshape(total_q, Hq, 128)
     # Blocks past a request's own extent stay -inf, so build the padded oracle.
     ref = torch.full((Hq, mkt, total_q), -float("inf"), dtype=torch.float32)
     for b, sk in enumerate(seqs_k):
